Extracting paywall object to its own file
